Tambah Fitur OS CLS dan merapikan kode

+++ M10_5230411296_Hafizh Habiibi Lubis/M10_5230411296_Hafizh Habiibi Lubis.py
@@ -1,5 +1,6 @@
 import mysql.connector
 import uuid
+import os
 import datetime as dt
 from tabulate import tabulate
 
@@ -16,7 +17,7 @@ cur = conn.cursor()
 # # membuat database
 # cur.execute("CREATE DATABASE penjualan")
 
-# # membuat tabel pegawai
+# membuat tabel pegawai
 # cur.execute("""CREATE TABLE Pegawai (
 #             NIK CHAR(5) NOT NULL PRIMARY KEY,
 #             Nama_Pegawai VARCHAR(25),
@@ -32,20 +33,20 @@ cur = conn.cursor()
 # #membuat tabel transaksi
 # cur.execute("""CREATE TABLE transaksi (
 #             No_Transaksi VARCHAR(10),
-#             Tanggal_Transaksi DATE,
-#             Kode_Produk VARCHAR(10),
+#             Tanggal_Transaksi DATETIME,
+#             Kode_Produk CHAR(5),
 #             Jumlah_Produk INT(2),
 #             PRIMARY KEY (No_Transaksi, Kode_Produk),
-#             FOREIGN KEY (Kode_Produk) REFERENCES produk(Kode_Produk))""")
+#             FOREIGN KEY (Kode_Produk) REFERENCES produk(Kode_Produk) ON DELETE CASCADE)""")
 
 # # membuat tabel struk membuat composite key 
 # cur.execute("""CREATE TABLE Struk (
 #             No_Struk CHAR(5) NOT NULL PRIMARY KEY,
-#             NIK CHAR(5) NOT NULL,
+#             NIK CHAR(5),  -- NIK bisa NULL
 #             No_Transaksi CHAR(7) NOT NULL,
 #             Total_Harga INT(10),
-#             FOREIGN KEY (NIK) REFERENCES pegawai (NIK),
-#             FOREIGN KEY (No_Transaksi) REFERENCES transaksi(No_Transaksi))""")
+#             FOREIGN KEY (NIK) REFERENCES Pegawai(NIK) ON DELETE SET NULL,
+#             FOREIGN KEY (No_Transaksi) REFERENCES Transaksi(No_Transaksi))""")
 
 class Pegawai:
     def __init__(self, NIK, nama_pegawai, alamat_pegawai):
@@ -54,9 +55,17 @@ class Pegawai:
         self.alamat_pegawai = alamat_pegawai
 
     def tambah_pegawai(self):
-        cur.execute("""INSERT INTO pegawai (NIK, Nama_Pegawai, Alamat_Pegawai) VALUES (%s, %s, %s)""", 
-                    (self._NIK, self.nama_pegawai, self.alamat_pegawai))
-        conn.commit()
+        try:
+            cur.execute("""INSERT INTO pegawai (NIK, Nama_Pegawai, Alamat_Pegawai) VALUES (%s, %s, %s)""", 
+                        (self._NIK, self.nama_pegawai, self.alamat_pegawai))
+            conn.commit()
+
+        except mysql.connector.errors.IntegrityError as e:
+
+            if "Duplicate entry" in str(e):
+                print(f"PEGAWAI DENGAN NIK {self._NIK} SUDAH ADA!")
+            else:
+                print(f"ERROR {e}")
 
     @staticmethod
     def lihat_pegawai():
@@ -73,6 +82,12 @@ class Pegawai:
     def hapus_pegawai(NIK):
         cur.execute("DELETE FROM pegawai WHERE NIK = %s", (NIK,))
         conn.commit()
+
+    @staticmethod
+    def cari_pegawai(NIK):
+        cur.execute("SELECT * FROM pegawai WHERE NIK = %s", (NIK,))
+        result = cur.fetchone()
+        return result is not None
 
 class Produk:
     def __init__(self, nama_produk, jenis_produk, harga):
@@ -103,13 +118,19 @@ class Produk:
         cur.execute("DELETE FROM produk WHERE Kode_Produk = %s", (kode_produk,))
         conn.commit()
 
+    @staticmethod
+    def cari_produk(kode_produk):
+        cur.execute("SELECT * FROM produk WHERE Kode_Produk = %s", (kode_produk,))
+        result = cur.fetchone()
+        return result is not None
+    
     def get_kode_produk(self):
         return self._kode_produk
 
 class Transaksi:
     def __init__(self):
         self.no_transaksi = f"TR{str(uuid.uuid4().int)[:5]}"
-        self.tanggal_transaksi = dt.date.today().strftime("%Y-%m-%d")
+        self.tanggal_transaksi = dt.datetime.now().strftime("%Y-%m-%d %H:%M")
         self.produk_list = []  # List untuk menyimpan produk dan jumlah
         self.total_harga = 0
 
@@ -154,15 +175,17 @@ class Struk:
             return "Pegawai Tidak Ditemukan"
 
     def cetak_struk(self):
-        print("="*5 + "TOKO KUE BHARATA" + "="*5)
+        print("="*10 + "RUMAH MAKAN BHARATA" + "="*10)
         print(f"No Struk        : {self.no_struk}")
         print(f"No Transaksi    : {self.transaksi.no_transaksi}")
         for produk, jumlah in self.transaksi.produk_list:
+            print("-"*39)
             print(f"Nama Produk     : {produk.nama_produk}")
             print(f"Jumlah Produk   : {jumlah}")
+        print("-"*39)
         print(f"Total Harga     : Rp. {self.transaksi.total_harga}")
         print(f"Nama Pegawai    : {self.nama_pegawai}")
-        print("="*20)
+        print("="*39)
 
     def simpan_struk(self):
         cur.execute("""INSERT INTO struk (No_Struk, NIK, No_Transaksi, Total_Harga) VALUES (%s, %s, %s, %s)""",
@@ -171,6 +194,7 @@ class Struk:
 
 def menu_produk():
     while True:
+        os.system("cls")
         try:
             print("="*10 + "MENU PRODUK" + "="*10)
             print("1. Tambah Produk")
@@ -189,33 +213,56 @@ def menu_produk():
                 # objek produk
                 produk = Produk(nama_produk, jenis_produk, harga)
                 produk.tambah_produk()
+                print("Produk Berhasil Ditambahkan!")
+                os.system("pause")
 
             elif pilih == 2:
                 Produk.lihat_produk()
+                os.system("pause")
 
             elif pilih == 3:
+                Produk.lihat_produk()
+                print("\n")
                 kode_produk = str(input("Masukan Kode Produk Yang Akan Dihapus : "))
-                Produk.hapus_produk(kode_produk)
+
+                if Produk.cari_produk(kode_produk):
+                    Produk.hapus_produk(kode_produk)
+                    print(f"Produk Dengan Kode {kode_produk} Berhasil Dihapus!")
+                else:
+                    print(f"Produk Dengan Kode {kode_produk} Tidak Ditemukan!")
+                os.system("pause")
 
             elif pilih == 4:
+                Produk.lihat_produk()
+                print("\n")
                 kode_produk_ubah = str(input("Masukan Kode Produk Yang Akan Diubah : "))
-                nama_baru = str(input("Masukan Nama Baru Produk : "))
-                jenis_baru = str(input("Masukan Jenis Baru Produk : "))
-                harga_baru = int(input("Masukan Harga Baru Produk : "))
+                
+                if Produk.cari_produk(kode_produk_ubah):
+                    nama_baru = str(input("Masukan Nama Baru Produk : "))
+                    jenis_baru = str(input("Masukan Jenis Baru Produk : "))
+                    harga_baru = int(input("Masukan Harga Baru Produk : "))
 
-                Produk.edit_produk(kode_produk_ubah, nama_baru, jenis_baru, harga_baru)
+                    Produk.edit_produk(kode_produk_ubah, nama_baru, jenis_baru, harga_baru)
+                    print(f"Produk Dengan Kode {kode_produk_ubah} Berhasil Diubah!")
+
+                else:
+                    print(f"Produk Dengan Kode {kode_produk_ubah} Tidak Ditemukan!")
+                os.system("pause")
 
             elif pilih == 5:
                 return
             
             else:
                 print("PILIHAN MENU TIDAK TERSEDIA!!")
+                os.system("pause")
 
         except ValueError:
             print("INPUTAN SALAH SILAHKAN INPUT DENGAN ANGKA!!")
+            os.system("pause")
 
 def menu_pegawai():
     while True:
+        os.system("cls")
         try:
             print("="*10 + "MENU PEGAWAI" + "="*10)
             print("1. Tambah Pegawai")
@@ -234,31 +281,53 @@ def menu_pegawai():
                 # objek pegawai
                 pegawai = Pegawai(NIK, Nama, Alamat)
                 pegawai.tambah_pegawai()
+                print(f"Pegawai Dengan NIK {NIK} Berhasil Ditambahkan!")
+                os.system("pause")
 
             elif pilih == 2:
                 Pegawai.lihat_pegawai()
+                os.system("pause")
 
             elif pilih == 3:
+                Pegawai.lihat_pegawai()
+                print("\n")
                 NIK = str(input("Masukan NIK Pegawai Yang Ingin Dihapus : "))
-                Pegawai.hapus_pegawai(NIK)
+
+                if Pegawai.cari_pegawai(NIK):
+                    Pegawai.hapus_pegawai(NIK)
+                    print(f"Pegawai Dengan NIK {NIK} Berhasil Dihapus!")
+                else:
+                    print(f"Pegawai Dengan NIK {NIK} Tidak Ditemukan!")
+                os.system("pause")
 
             elif pilih == 4:
+                Pegawai.lihat_pegawai()
+                print("\n")
                 NIK = str(input("Masukan NIK Pegawai Yang Akan Diubah : "))
-                nama_baru = str(input("Masukan Nama Baru Pegawai : "))
-                alamat_baru = str(input("Masukan Alamat baru Pegawai : "))
-                Pegawai.edit_pegawai(nama_baru, alamat_baru, NIK)
+
+                if Pegawai.cari_pegawai(NIK):
+                    nama_baru = str(input("Masukan Nama Baru Pegawai : "))
+                    alamat_baru = str(input("Masukan Alamat baru Pegawai : "))
+                    Pegawai.edit_pegawai(nama_baru, alamat_baru, NIK)
+                    print(f"Pegawai Dengan NIK {NIK} Berhasil Dihapus!")
+                else:
+                    print(f"Pegawai Dengan NIK {NIK} Tidak Ditemukan!")
+                os.system("pause")
 
             elif pilih == 5:
                 return
 
             else:
                 print("PILIHAN MENU TIDAK TERSEDIA!!")
+                os.system("pause")
 
         except ValueError:
             print("INPUTAN SALAH SILAHKAN INPUT DENGAN ANGKA!!")
+            os.system("pause")
 
 def menu_transaksi():
     while True:
+        os.system("cls")
         try:
             print("="*10 + "MENU TRANSAKSI" + "="*10)
             print("1. Tambah Transaksi")
@@ -268,6 +337,8 @@ def menu_transaksi():
             pilih = int(input("Masukan Pilihan Menu : "))
 
             if pilih == 1:
+                Pegawai.lihat_pegawai()
+                print("\n")
                 NIK_pegawai = input("Masukkan NIK Pegawai yang Melakukan Transaksi: ")
 
                 # Mengecek apakah NIK pegawai valid
@@ -275,10 +346,13 @@ def menu_transaksi():
                 pegawai = cur.fetchone()
                 if not pegawai:
                     print("NIK Pegawai Tidak Ditemukan!")
+                    os.system("pause")
                     continue  #pegawai tidak ditemuksn
 
                 transaksi = Transaksi()
                 while True:
+                    Produk.lihat_produk()
+                    print("\n")
                     kode_produk = input("Masukan Kode Produk (atau tekan Enter untuk selesai): ")
                     if kode_produk.strip() == "":
                         break
@@ -295,37 +369,46 @@ def menu_transaksi():
                         transaksi.tambah_produk(produk, jumlah_produk)
                     else:
                         print(f"PRODUK DENGAN KODE {kode_produk} TIDAK DITEMUKAN!")
+                        os.system("pause")
 
                 if transaksi.produk_list:
                     transaksi.tambah_transaksi()
                     print("TRANSAKSI BERHASIL DITAMBAHKAN!")
 
                     # Membuat struk dan mencetaknya
-                    struk = Struk(transaksi=transaksi, NIK_pegawai=NIK_pegawai)  
-                    struk.cetak_struk() 
+                    struk = Struk(transaksi=transaksi, NIK_pegawai=NIK_pegawai)
+                    print("\n")
+                    struk.cetak_struk()
+                    print("\n")
                     struk.simpan_struk()  # simpan struk ke db
+                    os.system("pause")
 
                 else:
                     print("TIDAK ADA PRODUK YANG DITAMBAHKAN DALAM TRANSAKSI.")
+                    os.system("pause")
 
             elif pilih == 2:
                 cur.execute("SELECT * FROM transaksi")
                 list_transaksi = cur.fetchall()
                 print(tabulate(list_transaksi, headers=["No_Transaksi", "Tanggal_Transaksi", "Kode_Produk", "Jumlah_Produk"], tablefmt="pretty"))
+                os.system("pause")
 
             elif pilih == 3:
                 return
 
             else:
                 print("PILIHAN MENU TIDAK TERSEDIA!!")
+                os.system("pause")
 
         except ValueError:
             print("INPUTAN SALAH SILAHKAN INPUT DENGAN ANGKA!!")
+            os.system("pause")
 
 def main():
     while True:
+        os.system("cls")
         try:
-            print("="*10 + "TOKO KUE BHARATA" + "="*10)
+            print("="*10 + "RUMAH MAKAN BHARATA" + "="*10)
             print("1. Produk")
             print("2. Pegawai")
             print("3. Transaksi")
@@ -348,9 +431,11 @@ def main():
             
             else:
                 print("PILIHAN MENU TIDAK TERSEDIA!!")
+                os.system("pause")
             
         except ValueError:
             print("INPUTAN SALAH SILAHKAN INPUT DENGAN ANGKA!!")
+            os.system("pause")
 
 
 if __name__ == "__main__":
